Fix Qt 6.12 build
Q_ENUMS got deprecated.

(cherry picked from commit 398d2d9c9ac210a224d191d9f892ef37fdfc4be6)

+++ kaddressbook/plugins/mergelib/autotests/mergecontactselectlistwidgettest.cpp
@@ -27,7 +27,7 @@ void MergeContactSelectListWidgetTest::shouldHaveDefaultValue()
     auto listWidget = selectListWidget.findChild<QListWidget *>(QStringLiteral("listwidget"));
     QVERIFY(listWidget);
     QCOMPARE(selectListWidget.selectedContact(), -1);
-    QCOMPARE(selectListWidget.conflictType(), KABMergeContacts::MergeContacts::None);
+    QCOMPARE((int)selectListWidget.conflictType(), (int)KABMergeContacts::MergeContacts::None);
 }
 
 QTEST_MAIN(MergeContactSelectListWidgetTest)
